+++ tests/deprecated_api/test_remove_1-6.py
@@ -249,3 +249,19 @@ def test_v1_6_0_ddp_plugin_task_idx():
     plugin = DDPPlugin()
     with pytest.deprecated_call(match='Use `DDPPlugin.local_rank` instead'):
         _ = plugin.task_idx
+
+
+def test_v1_6_0_lightning_module_loaded_optimizer_states_dict():
+    from pytorch_lightning.core.lightning import warning_cache
+    model = BoringModel()
+    _ = model.loaded_optimizer_states_dict
+    assert any(
+        "The `LightningModule.loaded_optimizer_states_dict` property is deprecated in v1.4" in w for w in warning_cache
+    )
+    warning_cache.clear()
+
+    model.loaded_optimizer_states_dict = {}
+    assert any(
+        "The `LightningModule.loaded_optimizer_states_dict` property is deprecated in v1.4" in w for w in warning_cache
+    )
+    warning_cache.clear()